Create c17.v
Main issues: are all with respect to "design 2 of 4 - single stage pipeline at IPs/OPs" - the implemented schematic (screenshot 1 attached) only shows inputs for the Dff module and not from the NAND gates. The waveform (screenshot 2 attached) shows invalid outputs for N22 and N23 even after trying to declare variables so that N22 and N23 are the final output from the modified c17.v design.